+++ CHANGES.md
@@ -20,6 +20,7 @@ Release Notes.
 * Adapt Armeria's plugins to the latest version 1.22.x
 * Fix tomcat-10x-plugin and add test case to support tomcat7.x-8.x-9.x.
 * Fix thrift plugin generate duplicate traceid when `sendBase` error occurs
+* Support keep trace profiling when cross-thread.
 
 #### Documentation
 * Update docs of Tracing APIs, reorganize the API docs into six parts.